Stylecop fixes
