+++ ScriptedEvents/API/Features/Exceptions/DisabledScriptException.cs
@@ -4,11 +4,12 @@
 
     public class DisabledScriptException : Exception
     {
-        public string ScriptName { get; }
         public DisabledScriptException(string scriptName)
             : base($"The given script '{scriptName}' is disabled.")
         {
             ScriptName = scriptName;
         }
+
+        public string ScriptName { get; }
     }
 }
