+++ ScriptedEvents/Actions/NullAction.cs
@@ -6,14 +6,6 @@
     // Represents a line in a file that does not have any actions.
     public class NullAction : IAction, IHiddenAction, ICustomReadDisplay
     {
-        public string Name => "NULL";
-
-        public string[] Aliases => Array.Empty<string>();
-
-        public string[] Arguments { get; set; }
-
-        public string Type { get; }
-
         public NullAction()
         {
             Type = "UNKNOWN";
@@ -23,6 +15,14 @@
         {
             Type = type;
         }
+
+        public string Name => "NULL";
+
+        public string[] Aliases => Array.Empty<string>();
+
+        public string[] Arguments { get; set; }
+
+        public string Type { get; }
 
         public bool Read(out string display)
         {
