+++ ScriptedEvents/Actions/WaitUntilAction.cs
@@ -39,6 +39,7 @@
                     Log.Warn($"[WAITUNTIL] WaitUntil condition error: {response.Message}");
                     break;
                 }
+
                 yield return Timing.WaitForSeconds(1f);
             }
         }
