+++ ScriptedEvents/DemoScripts/ConditionSamples.cs
@@ -2,7 +2,6 @@
 {
     public class ConditionSamples : IDemoScript
     {
-
         public string FileName => "ConditionSamples";
         public string Contents => @"!-- DISABLE
 #Wait until 10 seconds after the round started.
