+++ ScriptedEvents/DemoScripts/IDemoScript.cs
@@ -3,6 +3,7 @@
     public interface IDemoScript
     {
         public string FileName { get; }
+
         public string Contents { get; }
     }
 }
