+++ ScriptedEvents/Structures/Argument.cs
@@ -57,6 +57,7 @@
                 {
                     return "Variable";
                 }
+
                 return Type.Name;
             }
         }
