+++ ScriptedEvents/Structures/InfectRule.cs
@@ -4,15 +4,15 @@
 
     public readonly struct InfectRule
     {
-        public RoleTypeId OldRole { get; }
-        public RoleTypeId NewRole { get; }
-        public bool MovePlayer { get; }
-
         public InfectRule(RoleTypeId oldRole, RoleTypeId newRole, bool movePlayer)
         {
             OldRole = oldRole;
             NewRole = newRole;
             MovePlayer = movePlayer;
         }
+
+        public RoleTypeId OldRole { get; }
+        public RoleTypeId NewRole { get; }
+        public bool MovePlayer { get; }
     }
 }
